ENH: Introduce base assistant.
