+++ src/darsia/__init__.py
@@ -58,4 +58,5 @@ from darsia.manager.analysisbase import *
 from darsia.manager.concentrationanalysisbase import *
 from darsia.manager.traceranalysis import *
 from darsia.manager.co2analysis import *
+from darsia.assistants.base_assistant import *
 from darsia.assistants.rotation_correction_assistant import *
