+++ src/darsia/assistants/rotation_correction_assistant.py
@@ -8,7 +8,7 @@ import numpy as np
 import darsia
 
 
-class RotationCorrectionAssistant:
+class RotationCorrectionAssistant(darsia.BaseAssistant):
     """Class defining an assistant to set up :mod:`darsia.RotationCorrection`."""
 
     def __init__(self, img: darsia.Image, **kwargs) -> None:
@@ -24,15 +24,14 @@ class RotationCorrectionAssistant:
                 verbosity (bool): flag controlling verbosity, default is False
 
         """
-        self.img = img
+        super().__init__(img, **kwargs)
+
+        # Initialize containers
         self.rotation_corrections = []
-        self.kwargs = kwargs
+        self._reset()
 
         # Determine the center of the image
         self.center = 0.5 * (self.img.origin + self.img.opposite_corner)
-
-        # Initialize anchor and src/dst points
-        self._reset()
 
     def __call__(self) -> None:
         """Call the assistant."""
@@ -63,6 +62,26 @@ class RotationCorrectionAssistant:
             print("anchor: {}".format(self.anchor))
             print("src: {}".format(self.src))
             print("dst: {}".format(self.dst))
+
+    def _print_instructions(self) -> None:
+        """Print instructions."""
+
+        print("\nWelcome to the rotation correction assistant.")
+        print("Instructions for determining rotation correction:")
+        print("1. Consider merely one of the three subplots at a time.")
+        print("2. Right click to add anchor point.")
+        print("3. Left click to add src points. Repeat left click to add dst points.")
+        print(
+            """4. Make sure to choose one anchor point, and at least two src and dst """
+            """points. Press 'enter' to finalize the subselection. The corresponding """
+            """rotation correction will be computed and stored."""
+        )
+        print("5. Do not close the figure yourself.\n")
+
+    def _setup_event_handler(self) -> None:
+        """Define events."""
+        self.fig.canvas.mpl_connect("button_press_event", self.on_mouse_click_3d)
+        self.fig.canvas.mpl_connect("key_press_event", self.on_key_press)
 
     def on_key_press(self, event) -> None:
         """Finalize selection if 'enter' is pressed, and reset containers if 'escape'
@@ -184,111 +203,6 @@ class RotationCorrectionAssistant:
                 "Cannot add points while in {} mode. Deactivate it first.".format(state)
             )
 
-    def _plot_3d(self) -> None:
-        # Setup figure
-        fig_2d, axs = plt.subplots(1, 3)
-        fig_2d.suptitle("2d side views")
-        _title = "rotation assistant"
-
-        # Setup event handler
-        self.fig = fig_2d
-        self.ax = axs
-        fig_2d.canvas.mpl_connect("button_press_event", self.on_mouse_click_3d)
-        fig_2d.canvas.mpl_connect("key_press_event", self.on_key_press)
-
-        # Print instructions
-        print("\nWelcome to the rotation correction assistant.")
-        print("Instructions for determining rotation correction:")
-        print("1. Consider merely one of the three subplots at a time.")
-        print("2. Right click to add anchor point.")
-        print("3. Left click to add src points. Repeat left click to add dst points.")
-        print(
-            """4. Make sure to choose one anchor point, and at least two src and dst """
-            """points. Press 'enter' to finalize the subselection. The corresponding """
-            """rotation correction will be computed and stored."""
-        )
-        print("5. Do not close the figure yourself.\n")
-
-        # Fetch bounding box
-        corners = np.vstack((self.img.origin, self.img.opposite_corner))
-        bbox = np.array([np.min(corners, axis=0), np.max(corners, axis=0)])
-
-        # Extract physical coordinates and flatten
-        matrix_indices = np.transpose(np.indices(self.img.shape[:3]).reshape((3, -1)))
-        coordinates = self.img.coordinatesystem.coordinate(matrix_indices)
-
-        # Extract values
-        assert not self.img.series
-        array = self.img.img
-        flat_array = array.reshape((1, -1))[0]
-
-        # Restrict to active voxels
-        threshold = self.kwargs.get("threshold", np.min(self.img.img))
-        relative = self.kwargs.get("relative", False)
-        if relative:
-            threshold = threshold * np.max(self.img.img)
-        active = flat_array > threshold
-
-        # Signal strength
-        alpha_min = 0.1
-        alpha = alpha_min + (
-            (1.0 - alpha_min)
-            * (flat_array - np.min(array))
-            / (np.max(array) - np.min(array))
-        )
-        scaling = self.kwargs.get("scaling", 1)
-        s = scaling * alpha
-
-        # xy-plane
-        axs[0].set_title(_title + " - x-y plane")
-        axs[0].scatter(
-            coordinates[active, 0],
-            coordinates[active, 1],
-            s=s[active],
-            alpha=alpha[active],
-            c=flat_array[active],
-            cmap="viridis",
-        )
-        axs[0].set_xlim(bbox[0, 0], bbox[1, 0])
-        axs[0].set_ylim(bbox[0, 1], bbox[1, 1])
-        axs[0].set_xlabel("x-axis")
-        axs[0].set_ylabel("y-axis")
-        axs[0].set_aspect("equal")
-
-        # xz-plane
-        axs[1].set_title(_title + " - x-z plane")
-        axs[1].scatter(
-            coordinates[active, 0],
-            coordinates[active, 2],
-            s=s[active],
-            alpha=alpha[active],
-            c=flat_array[active],
-            cmap="viridis",
-        )
-        axs[1].set_xlim(bbox[0, 0], bbox[1, 0])
-        axs[1].set_ylim(bbox[0, 2], bbox[1, 2])
-        axs[1].set_xlabel("x-axis")
-        axs[1].set_ylabel("z-axis")
-        axs[1].set_aspect("equal")
-
-        # yz-plane
-        axs[2].set_title(_title + " - y-z plane")
-        axs[2].scatter(
-            coordinates[active, 1],
-            coordinates[active, 2],
-            s=s[active],
-            alpha=alpha[active],
-            c=flat_array[active],
-            cmap="viridis",
-        )
-        axs[2].set_xlim(bbox[0, 1], bbox[1, 1])
-        axs[2].set_ylim(bbox[0, 2], bbox[1, 2])
-        axs[2].set_xlabel("y-axis")
-        axs[2].set_ylabel("z-axis")
-        axs[2].set_aspect("equal")
-
-        plt.show()
-
     def _finalize_selection(self) -> None:
         """Finalize selection and setup rotation correction."""
         if (
@@ -331,7 +245,7 @@ class RotationCorrectionAssistant:
                 """'enter' to finish."""
             )
 
-    def return_rotation_corrections(self) -> list[darsia.RotationCorrection]:
+    def return_result(self) -> list[darsia.RotationCorrection]:
         """Return rotation correction.
 
         Returns:
